08-microservices/account_service/requirements.txt
flask==2.3.3
requests==2.31.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-propagator-jaeger
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests